@@ -7,11 +7,68 @@
 #include "llvm/Support/TargetRegistry.h"
 #include "llvm/Support/TargetSelect.h"
 #include "llvm/Target/TargetMachine.h"
-#include "llvm/Transforms/Scalar.h"
-#include "llvm/IR/LegacyPassManager.h"
 
 #include <iostream>
-#include <stdlib.h>
+#include <unistd.h>
+
+class InputParser {
+public:
+    InputParser (int &argc, char **argv) {
+        if (argc == 1) {
+            show_usage();
+            exit(0);
+        }
+
+        for (int i = 1; i < argc; ++i) {
+            if (std::find(std::begin(supported_options), std::end(supported_options), std::string(argv[i]))
+                != std::end(supported_options) || std::string(argv[i-1]) == "-o" || i == 1) {
+                tokens.emplace_back(std::string(argv[i]));
+            } else {
+                std::cerr << "error: option '" << argv[i] << "' is not supported!" << std::endl;
+            }
+        }
+    }
+
+    void show_usage() {
+        std::cout << "USAGE: turnip2 <input> [options]" << std::endl
+                << "OPTIONS:" << std::endl
+                << "\t -g          generate source-level debug information" << std::endl
+                << "\t -emit-llvm  emit LLVM IR for source inputs" << std::endl
+                << "\t -o <file>   write output to <file>" << std::endl
+                << "\t -O          optimize code to reduce size and time of execution" << std::endl
+                << "\t -S          only run compilation steps" << std::endl;
+    }
+
+    const std::string& get_option(const std::string &option) const {
+        std::vector<std::string>::const_iterator itr;
+        itr =  std::find(this->tokens.begin(), this->tokens.end(), option);
+        if (itr != this->tokens.end() && ++itr != this->tokens.end()){
+            return *itr;
+        }
+
+        for (auto &&token : tokens) {
+            if (token.find(option) == 0) {
+                return *new std::string(token.substr(token.find_first_of(option.back())+1));
+            }
+        }
+
+        return *new std::string("");
+    }
+
+    bool option_exists(const std::string &option) const {
+        return std::find(std::cbegin(tokens), std::cend(tokens), option) != std::cend(tokens);
+    }
+
+private:
+    std::vector<std::string> tokens;
+    const std::vector<std::string> supported_options = {
+            "-g",
+            "-emit-llvm",
+            "-o",
+            "-O",
+            "-S"
+    };
+};
 
 void generateObject(Module *m, std::string &out) {
     InitializeNativeTarget();
@@ -24,7 +81,7 @@ void generateObject(Module *m, std::string &out) {
     std::string error;
     auto target = TargetRegistry::lookupTarget(targetTriple, error);
 
-    if (!target) {
+    if (target == nullptr) {
         errs() << error;
         return;
     }
@@ -39,7 +96,7 @@ void generateObject(Module *m, std::string &out) {
     m->setDataLayout(theTargetMachine->createDataLayout());
 
     std::error_code EC;
-    raw_fd_ostream dest(out + ".o", EC, sys::fs::F_None);
+    raw_fd_ostream dest((out.find('.') != std::string::npos) ? (out.substr(0, std::string(out).find_last_of('.')) + ".o") : (out + ".o"), EC, sys::fs::F_None);
 
     if (EC) {
         errs() << "Could not open file: " << EC.message();
@@ -57,10 +114,14 @@ void generateObject(Module *m, std::string &out) {
     pass.run(*m);
     dest.flush();
 
-    system(std::string("gcc " + out + ".o " + " -o " + out).c_str());
+#if defined(__linux__)
+    execl("/usr/bin/gcc", "/usr/bin/gcc", std::string((out.find('.') != std::string::npos) ? (out.substr(0, std::string(out).find_last_of('.')) + ".o") : (out + ".o")).c_str(), std::string("-o" + out).c_str(), NULL);
+#endif
+
 }
 
 int main(int argc, char **argv) {
+    InputParser params(argc, argv);
     std::ifstream in(argv[1]);
 
     char c;
@@ -68,28 +129,45 @@ int main(int argc, char **argv) {
 
     while (!in.eof()) {
         in.get(c);
-        code.push_back(c);
+        code.emplace_back(c);
     }
-    code.push_back(EOF);
+    code.emplace_back(EOF);
 
     try {
         Lexer *lexer = new Lexer;
         lexer->load(code);
 
         Parser *parser = new Parser(lexer);
-        Node *ast = parser->parse();
+        std::shared_ptr<Node> ast = parser->parse();
 
-        Generator *generator = new Generator;
+        bool optimize = params.option_exists("-O");
+        bool generateDI = params.option_exists("-g");
+        if (optimize && generateDI)
+            generateDI = false;
+
+        Generator *generator = new Generator(optimize, generateDI, argv[1]);
         generator->generate(ast);
 
-        generator->module->dump();
-
-
         std::string output = "a.out";
-        if (argc >= 3)
-            output = std::string(argv[2]);
+        if (!params.get_option("-o").empty()) {
+            output = params.get_option("-o");
+        }
 
-        generateObject(generator->module, output);
+        if (params.option_exists("-emit-llvm")) {
+            std::string file = std::string(argv[1]).substr(0, std::string(argv[1]).find_last_of('.')) + ".s";
+            std::error_code EC;
+            raw_fd_ostream dest(file, EC, sys::fs::F_None);
+            generator->module.get()->print(dest, nullptr);
+            dest.close();
+        }
+
+        if (!params.option_exists("-S")) {
+            generateObject(generator->module.get(), output);
+
+            if (generateDI) {
+                generator->dbuilder->finalize();
+            }
+        }
     }
     catch (const std::string &err) {
         std::cerr << "In file " << argv[1] << ":" << err << std::endl;
